Get before deleting Azure cloud resources

+++ pkg/provider/cloud/azure/security_group.go
@@ -206,6 +206,16 @@ func ensureSecurityGroup(ctx context.Context, clients *ClientSet, cloud kubermat
 }
 
 func deleteSecurityGroup(ctx context.Context, clients *ClientSet, cloud kubermaticv1.CloudSpec) error {
+	// We first do Get to check existence of the security group to see if its already gone or not.
+	// We could also directly call delete but the error response would need to be unpacked twice to get the correct error message.
+	res, err := clients.SecurityGroups.Get(ctx, cloud.Azure.ResourceGroup, cloud.Azure.SecurityGroup, "")
+	if err != nil {
+		return err
+	}
+	if isNotFound(res.Response) {
+		return nil
+	}
+
 	future, err := clients.SecurityGroups.Delete(ctx, cloud.Azure.ResourceGroup, cloud.Azure.SecurityGroup)
 	if err != nil {
 		return err
